requirements.txt
asgiref==3.2.7
astroid==2.4.0
beautifulsoup4==4.6.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-bootstrap4==1.1.1
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.2
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.8.1
webencodings==0.5.1
Willow==1.3
wrapt==1.12.1
